requirements.txt
torch
torchvision
datasets
nltk
pillow
tqdm
requests
